Add individual regularization to  _graph_lasso.py
In the paper by Friedman et al. (2008), there is a REMARK 2.1, which states that the objective function to be optimized can be modified to allow for a matrix of penalty values, rather than a scalar value. This has been implemented.
However, linear_model._cd_fast.pyx still has to be updated to allow for vectorized alpha input in enet_coordinate_descent_gram